Install react-router-dom and node-sass packages, reorganize react file tree, add navbar and addAdmin components
