+++ react-frontend/src/components/NavBar.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+import './scss/NavBar.scss'
+
+class NavBar extends Component {
+  render() {
+    return (
+      <div className='NavBar'>
+        NavBar
+      </div>
+    );
+  }
+}
+
+export default NavBar;
