+++ react-frontend/src/components/scss/BtnGeneric.scss
@@ -1,0 +1,5 @@
+.BtnGeneric {
+  color: white;
+  background: linear-gradient(to right, rgba(101, 213, 209, 1), rgba(167, 234, 201, 1));
+  text-transform: uppercase;
+}
