+++ react-frontend/src/components/scss/NavBar.scss
@@ -1,0 +1,3 @@
+.NavBar {
+  width: 100vw;
+}
