+++ react-frontend/src/pages/scss/AddAdmin.scss
@@ -1,0 +1,4 @@
+.AddAdmin {
+  background: white;
+  color: black;
+}
