+++ react-frontend/src/pages/scss/App.scss
@@ -1,0 +1,4 @@
+.App {
+  text-align: center;
+}
+
